bug 1029: Updated StandingsJSON2016 to use the new ProblemSummaryInfo methods to get pending and judged run counts.
Unfortunately, those methods don't seem to produce the correct values so
the Scoreboard results are still incorrect (see comments in bug 1028).

Also added IInternalController parameter to the StandingsJSON2016
createJSON() method, in order to be able to access the current log.
This in turn required also updating ScoreboardService and
StandingsJSON2016Test.