bug 1052: added support for /language service to match the CLI spec at https://clics.ecs.baylor.edu/index.php/Draft_CCS_REST_interface#GET_baseurl.2Flanguages. Tested with zero, 1, 2 and 3 languages; verified output matches spec.
Note: the spec says the ID value for each language is to be "an
integer".  The implementation uses the Long value which is part of the
toString() of the ElementID for the language as this integer.  Since the
value of a Long is an integer, this technically meets the spec...
although I'm betting there are some who will argue that an integer must
fit into an "int" (which I would argue is incorrect).