bug 1055: changed ContestInformation to use GregorianCalendar instead of Date for scheduledStartTime.
Actually, MOST of this change was made in an earlier commit, but I found
some places where it was still referring to the (otherwise unused) Date
startDate.   Since there are external classes which both set and get the
scheduledStartTime and expect to use Date objects to do that, also
updated the getter/setter to convert internally between Date and
GregorianCalendar.