bug 1117: added create of problemset.yaml to export of contets.yaml
add unit tests, added javadoc and constants
This is task 1 - create problemset.yaml, at this time there
are now duplicates of problems in problemset.yaml and contest.yaml.
At this time the import for pc2 only imports contest.yaml so
there will be no duplicates added.   When task 2 and task 3 
are done then the problem set section will only be in problemset.yaml