bug 1074: Modified MSTOVPane Judge's output error message for clarity.
Previously, when the judge selected "View Team Output" and the system
couldn't find an output file for the run, it output a cryptic message
"Unable to find data for Table Cell <row>,<col>".  This caused judges in
the most recent CSUS contest to go "huh?"

Now, it says "No Team Output file found".