bug 1147: Fixed autostart under the condition that Settins were changed.
As Troy indicated, there are three identified conditions under which a
scheduled AutoStart fails:

- when loading Yaml containing a start time
- when making an update to Contest Settings via the Admin when a
Scheduled Start Time has been loaded from Yaml
- when restarting a Server where a Scheduled Start Time has previously
been set but the time has not yet arrived when the Server is restarted.

All of these appear to be related to failure to update the "hidden
isAutoStart" flag (the flag that remains in the system even though we
removed the corresponding checkbox from the GUI).

This checkin fixes the first of those; it's precisely Troy's one-line
fix in ContestInformationPane.