bug 1159: checking in updates to EditProblemPane and related changes.
This checkin is known to break several things (but only on this branch;
it shouldn't affect the master branch at all).  Known breakages are in
api.ServerConnection, ContestYAMLLoader and ContestSnakeYAMLLoader,
EditProblemPaneNew, and a number of JUnits in the test package.  

I'm working on fixing those things (although a group discussion is
probably going to be required).  The main reason for checking this in is
so that anyone working on bugs 1153 and/or its dependencies can see the
work I've been doing -- in particular, take a look at the new
"Validator" pane on the Admin.