bug 1074: updated Admin Guide .doc and .pdf.
Completely rewrote the Chapter 2 "Checklist"; it is now a much more
complete description of "how to get going".  (I did this in response to
the repeated comments we've gotten from people claiming they've read the
manual and can't figure it out -- I think it was just too daunting and I
hope this will help).

Also Updated some screen shots to match V9.4, as well as documented a
few of the 9.4 changes (still have many to go...)