bug 1202: Fixed the situation where "TLE" submissions were incorrectly marked as "CE".
Updated the default "TLE" string in InternalController to match the
string returned by Executable (and the strings defined in
samps/reject*ini).

Changed default judgement when there is a mismatch between the Judgement
string and the existing Judgement messages from "CE" (element [1] of the
Judgements array) to element [2] of the Judgements array (which should
always be a non-variable-score "No" judgment).  (Files where this
occurred included SelectJudgementPaneNew, JudgementUtilities, and
AutoJudgingMonitor.)

Added a JUnit test.