bug 1159: revert change of bad code in SerializedFile.
The SerializedFile class now once again computes incorrect SHA1 sums,
the way it has for years.
The correction appears as a comment in method
SerializedFile.generateSHA1(), but it has been commented-out in order to
keep the class operating the way it used to.
A separate checkin with changes to handle both backwards-compatible bad
SHA1 codes and correct SHA256 codes will be considered at some point.