bug 1233: added support in RunList for site-specific starting RunIDs.
RunList now looks in pc2v9.ini (if present) and checks for a property
"server.baseRunNumber".  If the property is found and is a legal
positive decimal number then that number becomes the RunID for the first
run AT THE CURRENT SITE.

If different sites have different baseRunNumbers in their pc2v9.ini
files, their runs will start at different ID values.  It is the user's
responsibility to insure that baseRunNumbers do not overlap between
sites.  (Meaning, this needs to be documented in the Admin Guide.)