bug 1238: fixed an issue with failure to obtain a run which was "Being_Judged".
Previously, if the run came back from the server in a packet addressed
to the Feeder client (as opposed to a broadcast packet addressed to
everybody), the code was assigning the fetched run to a variable which
was local in scope to the RunListener.  Now it properly assigns the run
to a class-wide variable which the fetchRun() method can see.