bug 1160: moved the invocation of enableUpdateButton() in the DefineInputValidatorPane to invokeLater().
This was necessary because the method is called from the Swing thread
when a keystroke occurs in a textfield; the enableUpdateButton() method
is grabbing the value from the textfield but the textfield value wasn't
being updated by the Swing thread until after the key handler (which
also runs on the Swing thread) finished -- a classic synchronization
issue.  Moving the call to enableUpdateButton() to invokeLater() solves
this problem.