bug 1160: First working version of reorganized InputValidatorPane.
The InputValidatorPane, and each of its subpanels, is now a separate
JPanePlugin class.

This checkin also fixes a problem with editing an Input Validator
Program name by replacing it with blanks; this was causing a Problem to
be marked "has no Input Validator" even if it DOES have an Input
Validator Command line.  (A problem can have an Input Validator by
virtue of having only an Input Validator Command line...)