bug 936: Changed default to "do NOT stop execution on first failed test case.
I believe that we eventually want the default to be DO stop on 1st
failed test case, but it seems to be safer at the moment to default to
the condition which is the same as the previous (unmodified) version of
the code.  We can change it later if desired.