bug 1283: added preventive checks for null results.
Under some (currently unreproduceable) conditions the SwingWorker thread
that runs the Input Validator is apparently generating null results (see
the initial bug description).  So far nobody has been able to reproduce
the bug reliably (but it's known to exist; it has happened repeatedly
and was witnessed by at least three people).

To prevent throwing an NPE when this happens, I added code which checks
for a Null result from the Swingworker thread and skips processing it
(and also logs that it happened).

Ran the Input Validator and verified it still works; ran all 787 JUnit
tests (all passed).