bug 1256: corrected logic in saving Validator results.
Previously, the number of validator output and stderr result files
generated was assumed to be based on the number of judge's date files or
answer files.  That's not true if a problem is configured with no data
or answer files but has a custom validator which does something outside
the normal flow (e.g. because the judge wants to do external manual
judging).