bug 1268: Added code to display unexecuted test cases when Viewoutputs & Data is pushed.
Previously, the addAnyUnexecutedTestCases() method had a check at the
beginning for stopOnFirstFailure(); if this flag was false then it
assumed all test cases must have been executed. That is only true if the
Execute button was pushed; if the FIRST thing the use did was select
"View Outputs & Data", no runs had been executed whether or not
stopOnFirstFailure() was true.  Removed that check.
Also added code in SelectJudgementPane to set data into the frame in the
case of pushing ViewOutputsAndData.