bug 1425:  added check for existence of at least one judge's answer file before enabling compare button.
Solves the issue identified in Comment 3.  Passes all unit tests. Works
correctly for the tests identified in Comment 1 and comment 3 and also
with the samps/sumithello contest.

Note: the issue was that the Compare button was being enabled when there
were more than zero rows selected; the patch was to also test for the
existence of a judge's output (answer) file as well.  Note that this
presumes a problem will be configured either with NO judge's answer
files or with a judge's answer file for every test case.