bug 1448: fixed a problem with redisplay of run source.
Specifically, once you had requested to View Source for a run, all
subsequent attempts to View Source for that same run failed.  This was
because the Controller.fetchRun() method checks to see if the RunFiles
for the run are already present in the local contest model, and if so
skips fetching them from the server.  This is reasonable, but the code
wasn't allowing for this -- it was ALWAYS waiting for a response from
the server.  Now it only requests a FETCH_RUN from the server if the
RunFiles for the selected run are not already in the local model.