bug 1471: reinserting updated copies of Admin Guide.
The Admin Guide was updated on Dec 8, as mentioned in commit
a54dfdb68701054a2992c218c6b1151605e9f797, but somehow on Dec 26
something overwrote the updated version.  This commit restores the
update, which has changes as mentioned in that commit:

===========
bug 1471: updated Admin Guide to explain about the need for unique team
ids.

Specifically, added an explanatory paragraph to Appendix L, "PC2 Event
Feed".  Also changed the name of the appendix to include "XML", updated
the text to explain the current relationship between the CLICS XML and
JSON event feeds.  

Also removed several occurences of "ACM" throughout the text.
============

In addition a few minor rewordings were included in the new version
uploaded in this commit.