bug 1206: Refactored various shadow classes.
Added fleshed-out implementations for several classes based on the
descriptions in related bugs bug 1624 through bug 1628.  (However, no
claim is made that any one of the implementations is yet complete; work
remains to be done.)

I'm pushing this to the "develop" branch, under bug 1206, because it
comprises work spread out over that "umbrella bug"; however, recommend
that further work on the individual classes be done under the relevant
"feature bugs".