bug 1622: Split ShadowSettingsPane into a separate class.
Originally, the code to create the Shadow Mode Settings pane was
embedded within ContestInformationPane (the GUI for Contest Settings).
However, the Shadow Settings need to be displayed in other places -- for
example, in the ShadowControlPane on the EF Client.  Separating out
ShadowSettingsPane into its own class makes this easy to do.