bug 1624: revert back to Sting representation for URL.
It turned out to be VERY complex to use a URL class for holding the URL
representation, due to the fact that every single keystroke in the URL
textbox is captured and validated.  Rather than totally restructure the
keyboard input validation mechanism (which is spread throughout the
system in a similar fashion), it is much easier to maintain the URL
representation as a String, and then convert it to an actual URL
instance at the time the URL is actually passed to the classes behind
the ShadowController facade (such as the RemoteRunMonitor).