bug 1637: changed Thread sleep time calculation.
Previously, the delay between successive event outputs from the Mock
Adapter was configurable only by whole seconds.  If this value was
specified as zero when passed to PacedFileInputStream (causing no
delay), the Mock overwhelms the rest of the system.  If this value is
greater than zero but even as low as 1 second, a contest like ACPC 2019
with its 2800+ events takes over 50 minutes to submit all the contest
events.

Now, instead of multiplying the sleepSeconds by 1000 to determine sleep
delay msec, it is multiplied by 20.  If the Mock then specifies a sleep
time of "1", that inserts 1*20msec between each event.   So the meaning
of "sleep time" is now "the number of 20msec delays between event
output.