bug 1637: changed to use Validator Result if such exists.
Previously, the JSONTool by default inserted the value returned by
getJudgementType(judgement) as the value of "judgement_type_id" in the
JSON for a judgement.  Now, it first checks whether there was a
Validator Result for the submission; if so, it attempts to use the
Validator Result to obtain a CLICS judgement type to insert into the
JSON.  (If there was no Validator Result, or if the Validator Result
couldn't be used to obtain a CLICS judgement, it still falls back to
using the submission judgement.)

The reason for doing this is that PC2 assigns a default judgement type
of "RTE" to any submission that is neither an AC nor a CE; defaulting to
using the judgement type results in many judgements being displayed as
"RTE" when in fact a Validator was used and returned some other
judgement.

Also added "compile error" to strings indicating "CE".