bug 1637: Updated to use Validator Results if such exist.
Previously, by default the Shadow Controller used the value returned by
getJudgementType(judgement) to determine the acronym for a judgement.
Now, it first checks whether there was a Validator Result for the
submission; if so, it attempts to use the
Validator Result to obtain a judgement acronym.  (If there was no
Validator Result, or if the Validator Result
couldn't be used to obtain a judgement, it still falls back to
using the submission judgement.)

The reason for doing this is that PC2 assigns a default judgement type
of "RTE" to any submission that is neither an AC nor a CE; defaulting to
using the judgement type results in many judgements being displayed as
"RTE" when in fact a Validator was used and returned some other
judgement.

Also added a source code boolean indicating whether results should
automatically be translated to CLICS-equivalent acronyms, and a method
to do this conversion if requested.  Default is "true"; this variable
should be changed to be a user-controllable attribute.

Also deleted unused variables like "listening".