bug 1637: updated ExtractPlaybackLoadFilesReport to report CLICS judgements.
Previously this class, aka ExtractReplayRunsReport, output "judgements"
which were derived by looking at the "judgementID" assigned to a
judgement.  This is completely deceptive since the default "judgementID"
assigned to runs is "No - Runtime Error", and such a judgement is never
updated if the validator -- or a judge -- assigns a new judgement value.

The most appropriate representation of a judgement is the content of the
"validator string" in the judgement (if it exists). Most PC2 classes
(e.g. the "Runs Pane") display judgements based on the "validator
string" if it exists, falling back to the "judgementID" toString() only
of there is no validator string.

Using the judgementID rather than the validator string (if it exists) is
very misleading and causes problems with other functions (such as Shadow
mode) which are expecting validator result strings.  This change causes
the ExtractReplayRuns report output to contain validator strings (like
most other displays in PC2) rather than the judgementID which is
assigned by default.