bug 1637: changed convertToJSON(Account) to use team displayName.
Previously it was putting the value returned by getTeamName() into the
JSON as the name of the team; the problem is that this value is not
always up to date (see bug 1648).  Now it puts the value returned by
getDisplayName() into the JSON; this value is correctly updated when a
team account is edited.