bug 1652: fixed faulty zip file buffering in RemoteContestAPIAdapter.
Previously the code was using a
com.sun.xml.internal.messaging.saaj.util.ByteOutputStream.ByteOutputStream()
to collect bytes from a zipfile, then calling getBytes() on that object
to get the collected bytes.  This class always returns a fixed-size
buffer (default size 1024), which results in null-padding at the end of
the resulting IFile.

Now the code uses a java.io.ByteArrayOutputStream for buffering; this
class has the ability to return only the valid contents of the buffer,
with no padding.