bug 1625: added ability for shadow to process only selected submissions.
Previously if you started a Shadow Controller, the Remote Event Feed
Monitor would process every submission that was sent by the Primary CCS.
This caused a lot of delay and difficulty when using something like the
PacNW (with 604 submissions) as the Primary.

Now, there is a source code variable "respectSubmissionFilter" in
RemoteEventFeedMonitor.  The default value is false, which means it will
continue to process all submissions.  If the variable is set true, the
RemoteEventFeed Monitor will examine the String[] submissionFilterIDs
and only process submissions whose IDs are listed (in String form) in
that array.

Future work could include the ability to set this variable from, say,
the [shadow] section in pc2v9.ini.