bug 1655: replace catch fall-through with explicit return.
Previously, the code was written to ignore exceptions, and the intent
was that this condition should return false.  However, the catch block
was written as a "fall-through" to a return statement. If a subsequent
change is made to the code, falling-through might no longer produce the
intended return response.  This change simply adds "return false" in the
catch, where the intent is that "any exception should return false".