bug 1625: added code to skip submissions from the remote if already present in the shadow.
This condition could happen if the Shadow is restarted reading the event
feed from the beginning; we don't want to insert into the Shadow server
submissions that were already inserted by the earlier reading of the
event feed.

Also updated, then commented out, the code which uses the submission
event "files:[{href:path... }] element to determine the URL from which
to fetch submissions; that code can be restored when the Primary CCS URL
is split into "BaseURL" and "ContestIDPath".