bug 1625: restructured getRemoteSubmissionFiles().
Previously there were two methods, getRemoteSubmissionFiles(String) and
getRemoteSubmissionFilesNew(String).  Converted these into a set of
overloaded methods getRemoteSubmissionFiles(String/URL), with the first
one accepting a String "submissionID" and converting it to a default URL
and invoking the second one with that URL.

Also added JavaDoc for multiple methods.