bug 1668: made validator phase ExecutionTimer local to method.
This avoids a small timing window where the execute phase can terminate,
but the TLE-Timer task to kill the execute phase process has already
been scheduled (but hasn't completed its kill actions yet).  In this
scenario the validator can start running, using a new ExecuteTimer but
which (in the previous code) is ASSIGNED TO THE SINGLE GLOBAL
"executionTimer" VARIABLE. This causes the TLE-Timer to call
stopIOCollectors() on the *validator's* execution timer, which chokes
the validation phase (throws exceptions).