issue #49: Reuse any existing ShadowCompareFrame instead of creating a new one.
Also arranged for disposal of any existing ShadowCompareFrame when
shadowing is stopped (because restarting will need a NEW frame because
there will be a new ShadowController).

Also moved the management of the Compare button enabling/disabling into
the updateGUI() method (where it probably belonged to start with). Now
each time updateGUI() is called it enables/disables the Compare button
based on whether shadowing is currently ongoing (it makes no sense to
"compare" when there is no shadowing happening).