i64: Added support for userlogo/banner files in the login screen.
If the user puts a file "logo.{png,jpg}" in the /images folder,
then the current CSUS logo will be displayed in scaled-down form
in the upper left and the user's logo will be displayed as the
main logo image.  The user logo is scaled to the dimensions
of the original CSUS logo.

Also added a PC2 logo to Login screen; the PC2 logo image file
is verified using a checksum (as was already being done for the
CSUS and ICPC logo files).

Also added support for user-provided "banner.{png,jpg}" file;
if this file is present in /images then it will be used as the
bottom banner instead of the existing ICPC banner. The user
banner file is scaled to fit within the available bottom-panel
space.

Also updated the "Name" label on the Login screen to "Team Login Name".
i64: correct layout spacing of password title label.

Also updated PC2logo.ico to use the new V3 version of the PC2 balloon
logo.

Note: during development, code was added to place "line borders" around
all login screen components, which was useful to determine where various
images and other components were laid out.  All such "line border" code
has been commented out, but left in the code in case it might be useful
in the future.