i64: this is a squash of 4 commits (see comment details)
======================

i64: This is a squash of 7 different commits (see comment details).

-----------
i64: corrected GUI login name label ("Team Login Name"->"Login Name")

Also added log creation in main(); this is needed if the LoginFrame
class is run stand-alone (normally the class is invoked after
InternalController has established the log, but this doesn't happen if
LoginFrame.main() is invoked).

-----------
i64: fixed scaling of PC2logo.

Previously the PC2logo on the login screen was set to be 10% of the size
of the PC2Logo.png file which is read from a file.  This won't work if
the logo is later changed to a much larger resolution; the logo on the
login screen needs to always fit into the panel where it appears (but
the size of this panel isn't known in advance).

-----------
i64: rewrote FrameUtilities.verifyImage(); see extended comment.

Fixed a bug in verifyImage() which was causing timing-variable SHA
checksum computations (an error pointed out by Tim - thanks!)  With this
fix, the need for multiple "different SHA values" for a given image file
was eliminated, reducing code complexity significantly.  (It makes no
sense that a given image file should have to have multiple different SHA
checksums; the whole point of SHA is to produce a SINGLE (and -almost-
certainly unique) checksum. This fix removed the code that did that.

-----------
i64: renamed method in FrameUtilities; these files changed as a result.

-----------
i64: added icpc logo to the set of files inserted into pc2.jar/images.

The checksum for this file is already recognized in
FrameUtilities.loadAndVerifyImageFile(); this will make it easy to add
use of the ICPC Logo to future display updates.

-----------
i64: updated Admin Guide to reflect creation of "images" folder.

-----------
i64: updated Admin Guide r.e. copying customization images to all
clients.

======================

i64: fix typo in updated createVersionAndJar.xml

======================

i64: added comments listing hex SHA1 values for files in /images.