i107: Updated Admin Guide and WhatsNew, as follows:
This is a single commit containing multiple doc updates that were
originally pushed in a series of separate commits to PR 109 under Issue
107.  Those changes included:

- Listing WTI, Frozen Scoreboard, Customizable GUIS, and Shadow support
in WhatsNew

- Adding Appendices Q (GUI Customization) and R (Shadow Mode) to Admin
Guide

- Updating Admin Guide Appendix N (Team Clients) with WTI info

- Updating Admin Guide Ch. 10 (Scoreboards) with info on public/private
HTML and also with details on using XSL to transform XML.

- Minor updates throughout Admin Guide for compatibility with the above
changes.

The updates contained in this commit are exactly the combination of all
the above separate commits.  The single difference is that this commit
does not include the update to javadoc in the
DefaultScoringAlgorithm.java file that was pushed with the original set
of commits.