i75: added new ScoreboardChangeListener to ContestController.  Effect:
Any change in the contest state which can result in a change in the
standings (such as runs being submitted, judged, edited, or deleted, or
a Configuration change such as marking a problem or account deleted)
will invoke the new ScoreboardChangeListener in the ContestController.
The ScoreboardChangeListener sets the "wtiStandingsAreCurrent" flag in
the ContestController to false, meaning that the next time a client
requests the standings the ContestController will discard its cached
copy and instead invoke its DefaultScoringAlgorithm instance to obtain
updated standings (which it returns, and caches).

Added accessors for wtiServerStandingsAreCurrent so the
ScoreboardChangeListener can access (set) its value.

Also made minor code rearrangements for readability, and commented out
some debugging logging.