i75: add pc2v9 project to classpath (buildpath) of WTI project. Reason:
The WTI project contains a pc2.jar file in its WebContent libraries.  If
changes are made in the pc2v9 project to support WTI (for example,
updates to API classes, as are needed for the WTI Scoreboard being
developed under this bug), those changes are not visible to the WTI
project because they haven't been applied to the (old) pc2.jar in the
libraries.  Adding the pc2v9 project to the WTI project classpath (and
moving it up above "WebContent" in the "Order and Imports" BuildPath
configuration in Eclipse) makes changes in pc2v9 visible to WTI.