i118: updated MultiDataSetPane to match InputValidator changes. Specifically:
When new data files are loaded, the MultiDataSet pane checks to see if
an Input Validator has been defined, and if so prompts asking whether
the Input Validator should be run on the new files.  Previously, it
displayed a message saying "results will be displayed on the Input
Validator tab";  that's no longer true (results are displayed in a
separate InputValidationResultFrame).  Message was changed.

Also, a TODO note was added indicating that currently the
MultiDataSetPane *always* invokes "runCustomInputValidator()", even if
the "selected validator" is VIVA.  This needs to be fixed...