i177: update getOSName() to use browser's navigator.userAgent.  Also:
Commented-out references to numerous tried (and failed) alternatives for
getting the OS name from the underlying OS.  (I don't think you can do
that; browser javascript code is constrained to remain inside the
browser's sandbox, isn't it?)   That's why the decision was made to use
the navigator.userAgent; it IS accessible to javascript.

Fixes #177.