i259: restructure getJudgementComparisonInfo().  Specifically,
- changed the sequence of map-building steps so that it checks for null
values and is much more efficient about how it operates.  Now, instead
of getting PC2 runs, building a map from them, and then having to invert
that map to get the runs again, it simply uses the runs directly to
build the final required map.  It also now does this in a single place
instead of multiple places like before.  (It was the messed-up logic in
the previous sequence which was allowing the possibility of null entries
in the map, which in turn was causing NPEs when the Compare button was
pressed.)

- consolidated the decision about converting to "CLICS Big5" judgements,
and changed the way Big5 conversion is handled.

- eliminated some no-longer-needed methods.