i253: implement a toggle between status counts algorithms
Use Ctrl-Click on pane where status count labels are.  Ex. to
the left of No contact or to the right of READY label.
Add code to implement "new" requirement to count each action a
team has taken (instead of counting the team status colors).
See PR 264 for lots of detail.
Each time the display change is made it will be logged.
The display change will be reset each time a user logs in.