i273: updated runChanged() to ignore notices from other runs. Details:
the RunListener.runChanged() method is invoked on EVERY "run updated"
event, including (for example) those generated when a judge updates a
run.  Previously, the runChanged() method was always comparing the
updated run (from the RunEvent) with "the run we requested the server to
update".  This was throwing an NPE when we hadn't ever actually
requested the server to update a run.  Now the code checks to see if
we've previously requested an update, and resets the appropriate
variable to null once we get an update for our requested run.

Also added logging for various warning conditions such as getting a null
RunEvent or a null Run in a RunEvent.