i240: only allow RemoteEventFeedMonitor to submit when Feeder1. Detail:
As described in https://github.com/pc2ccs/pc2v9/issues/240, there are
times when it is important to have more than one "Shadow Controller"
running.  However it is also important that additional Shadow
Controllers not be allowed to forward submissions from the Remote CCS to
the PC2 server.  This commit adds a check to see what account a Shadow
client is currently logged in with; only Feeder1 is allowed to actually
forward (submit) a submission to the PC2 server.