i295: revert addition of File.separator in sandbox command line:
The evaluation of File.separator occurs at PC2 COMPILE time, which means
the separator character will be whatever is defined on the machine on
which the system is compiled -- but it needs to match whatever exists
on the RUNTIME machine.

Since the sandbox is only being developed for Linux (at present), this
was reverted back to a hardcoded "./" as the executable command line
prefix.