Fix null client ID used in exception report and spelling errors
At the point the exception is thrown, the original clientId variable was always null.  It should have been the client who requested the checkout of the run, namely, whoChangedRun.  The same problem existed for checking out Clarifications, and that too has been fixed (in the same module).
In addition, there was a wrong word used in the string text that was thrown:  "can not checked out" changed to "can not check out".