Create interface for Executable timer updates
An instance of IExecuteTimerFrame is now passed into Executable.  This interface deals with updating the status labels on whatever Frame is using Executable.  Change calls in Executable to pass along the interface to the ExecuteTimer.

Change ExecuteTImer to use  IExecuteTimerFrame and its methods instead of making direct calls to the label GUI components.

ExecuteTImerFrame is a general JFrame implementation of IExecuteTimerFrame that can be used if an existing JFrame is not available.

AutoJudgeStatusFrame changed to implement IExecuteTimerFrame and make use of it's existing labels and methods.

Change classes that use Executable to pass an instance of IExecuteTimerFrame or create an ExecuteTimerFrame if a JFrame is unavailable.