i480: fix incorrect hashtable mapping.  Specifically:
The original code was using "Run" as the key into the hashtable of "runs
which have at some point been added to the UnavailableRunsList".
However, neither class "Run" nor its parent class "Submission" defines a
"hashcode()" method -- which means different Objects representing the
same "Run" are going to have different hashcodes -- meaning in turn that
they will not map to the same hastable location if "Run" is used as the
key.

The fix, courtesy of John B., is to use a Run's "ElementId" as the key
into the hashtable.  ElementId DOES define hashcode(), and what it
returns is always the same hashcode for a given "Run" regardless of
whether it is a clone of the "original" Run object or not.  (That is,
the hashcode() returned by ElementId is a function of the Run "number"
(a.k.a. "run id"), which never changes.)