Make sure that fetchRun response is for correct run
It is possible (and likely in a busy contest) to get the incorrect run back when doing a View Source.  This is because no check was done to make sure that a runChanged event was for the specific run whose source was being requested.
Note that this fixes this bug on a single site contest.  On a multisite contest, the fix for bug #497 is also required or no response will be received from the remote site.