add the Dockerfile to rebuild pc2fromecs/jdk8hugo
with ubuntu22.04 which has the curl version that
supports --retry-all-errors

Also note the /root/node_modules with all the dependecies
for WTI-API are installed on this container.
It would be time safer to "cp -r /root/mode_modules projects/WTI-UI/"
effectively in the projects/WTI-API/buildWTI.xml if /root/node_modules
exists.

fixes #519