CI: Fix possible wrong selected run index for View Source
Move the code that obtains and validates the selected run index to the main thread.  Previously, the selected index was obtained in the spawned off View Source thread.  It is possible, on a slow client, that the user could have selected a different run before the thread had a chance to read the originally selected index.
The already validated selected index is now passed to the showSourceForSelectedRun(int) routine.