Fixes i543 - Event type not properly detected causes delays
Check for the event "type" after a map has been created of properties.  Different feeds format JSON differently, so it's better to parse the feed than to just look for patterns in the event line.
Only sleep on events that we process (submissions and judgements).  Use a flag to indicate if a delay is desired.
Add constant for how long to delay.