Indent method body
In preparation for creating a reconnect loop, simply add a benign 'if' and indent the whole block of the run() method.
There are no differences other than the if(keepRunning){ ... } and the block in the 'if' was indented.  This will make it easy to evaluate additional code changes without the potentional for a diff "jackpot".