Add shorthand method to clean up logging
Many places in the code want to print both a debugging messages to System.err AND log a message to the logger as well.  The code to do this is roughly 4-6 lines.  Instead, a method was created that checks if debugging is enabled then prints the message.  The message is always logged.
A method was also created for logging with an exception that was thrown.
Added record counter and remember last token for debug messages, and later restarting the feed at a different point.
This commit should have no PC2 logic changes.  It is mostly cosmetic and an effort to reduce module size.