Add reconnect on remote socket error (rewrite)
Change main code-flow into a loop that retries connecting to remote on socket error (IO Exception)
Add support for starting or restarting from last received event (lastToken).  This is useful if the connection is lost and we reconnect, or, if it is desired to start reading the remote feed at a different start point manually.  This involves changing the interface to take a new "token" parameter.  This may be set to null if the feed should be started from the beginning.
The MockContestAPIAdapter also had to be modified implement the getRemoteEventFeedStream(String) method.  The parameter is ignored.
Add member variables to control if connect retries are to be attempted, and how long to wait between retries. (reasonable defaults are set)
Add accessors and mutators for new members.  Also add accessor for number of records read (this will be used in the GUI interface).