Fix Issue #544 - duplicate submissions
Add map to keep track of submissions we have seen on the event feed.  This way, we can check if we have seen a submission before it has been entered into the system.
Set delay time for submissions and judgements to zero since the event feed speed issue #267 (pr #555) has been fixed.