Add execute substitution vars :infilename :ansfilename
Provide access to judge's file's real names by exposing them in :infilename and :ansfilename as substitution variables for the execute step.  This is useful for programs such as interactive validators so it can be determined which case a run fails on.  :infile and :ansfile are insufficient for this purpose since those always contain the same name for each test case.