Change JSON record generator for "runs" to use run case testNumber, fixes issue #621 (#622)
* Change JSON record generator for "runs" to use testcase number

Fixes #621

CLICS spec says the ordinal for "runs" test cases should start at 1, not 0.  This change simply adds 1 to the ordinal when the json string is generated.

* i621: changed ordinal for runs test case records to be the testCaseNumber

Instead of using the index into the entire vector of runTestCases for all judgements, use the specific test case number (getTestNumber()) as the ordinal for this judgement.