i564: correct getMaxFileSize log messages when problem is null:
an earlier version returned zero from getMaxFileSize() if the problem
was null; it was agreed to change this to instead return the global max
output size from ContestInformation and a PR to do this was pushed.
However, the log/error messages were not changed to reflect this, which
is what this commit does.