Add support for PC2 wall-time exceeded
PC2 sets a timer that is  (currently) 1 second longer the than CPU time limit.  This is to catch submissions that may sleep and not spend cycles.  The 1 second can be changed by changing the value of SANDBOX_EXTRA_KILLTIME_MS.  If the PC2 execute timer goes off, a TERMINATE (15) signal is sent to the pc2sandbox.sh script.  This is now caught and the script will kill the submission.  This is necessary since killing the pc2sandbox.sh would leave the submission running (it wouldln't normally be terminated since it's a child.).  The script catches the terminate signal from PC2 and kills off the submission and returns a new return code indicating that the wall time has been exceeded.  We simply count this as a time exceeded disposition.