Add support for check for OS compatibilities with sandboxes
Create scripts for LInux and Windows to check for sandbox support.  Windows currently returns an exit code of failure to indicate OS sandbox support is not present (yet).
Create OSCompatibilityUtilities class to see if problems need special OS support (sandboxes) and then execute script to test for OS presence of such support.
Add code in autojudge and judge models to make sure problem scan be judged on the client's OS.
Do not allow --nogui AJ's to start if executing problem submissions require special OS support that is not present.  Put detailed error in log if problems found.
Warn a GUI judge if some/all problems can not be judged on the client's OS with a nice message and a hint as to how to address it.